pyproject.toml
[project]
name = "to-do-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy-lite>=0.2.1",
    "pydantic>=2.12.5",
]
